% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{apply_greyscale_filter}
\alias{apply_greyscale_filter}
\title{Apply greyscale filter}
\usage{
apply_greyscale_filter(image)
}
\arguments{
\item{image}{The bitmap image.}
}
\value{
The filtered image.
}
\description{
This function takes the greyscale function and applies it to every pixel in the image.
}
