% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{apply_image_filter}
\alias{apply_image_filter}
\title{Apply image filter}
\usage{
apply_image_filter(image, fun, colour = c(255, 0, 0), ...)
}
\arguments{
\item{image}{The image to be filtered.}

\item{fun}{The filter of choice, greyscale or colour.}

\item{colour}{Filter colour of choice as a vector. Default is red c(255, 0, 0)}
}
\value{
The filtered image
}
\description{
This function applies the greyscale or two colour filter of choice.
}
