% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{apply_two_colour_filter}
\alias{apply_two_colour_filter}
\title{Apply two colour filter}
\usage{
apply_two_colour_filter(image, colour)
}
\arguments{
\item{image}{The bitmap image}

\item{colour}{The colour for the filter given as a vector, for example black c(0, 0, 0)}
}
\value{
The filtered image
}
\description{
This function takes the two colour filter and applies it to every pixel in the image.
}
