% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{greyscale_filter}
\alias{greyscale_filter}
\title{Greyscale filter}
\usage{
greyscale_filter(rgb)
}
\arguments{
\item{rgb}{the vector.}
}
\value{
greyscale vector.
}
\description{
This function takes a colour vector and returns a greyscale vector.
}
