% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{two_colour_filter}
\alias{two_colour_filter}
\title{Two colour filter}
\usage{
two_colour_filter(rgb, colour)
}
\arguments{
\item{rgb}{The colour vector.}

\item{colour}{The colour vector of choice for the filter}
}
\value{
A filtered colour vector with white colour and colour of choice
}
\description{
This function takes a colour vector and returns a new vector with the colour of choice
}
